debug

+++ include/StaticLinkList.h
@@ -45,7 +45,7 @@ protected:
 
     struct SNode : public Node
     {
-        void* operator new(unsigned int size, void* loc)
+        void* operator new(size_t size, void* loc)
         {
             (void)size;
             return loc;
